DEV: move gas to a separate module
